fix inventory button modify button

+++ dimabe_manufacturing/models/stock_production_lot.py
@@ -282,14 +282,15 @@ class StockProductionLot(models.Model):
     def fix_error_inventory(self):
         moves_with_problems = []
         move_state = self.env['stock.move'].search(
-            [('id', 'in', self.env['stock.move.line'].search([('product_id', '=', 2723)]).mapped('move_id').mapped('id'))]).mapped('state')
-        raise models.ValidationError(move_state)
-        for move in self.env['stock.move.line'].search([('product_id', '=', 2723)]):
-            if move.product_uom_qty > 0:
-                move.write({
-                    'state': 'cancel'
-                })
-                move.unlink()
+            [('id', 'in', self.env['stock.move.line'].search([('product_id', '=', 2723)]).mapped('move_id').mapped('id'))])
+        for stock in move_state:
+            if stock.state == 'done':
+                for move in stock.active_move_line_ids:
+                    if move.product_uom_qty > 0:
+                        move.write({
+                            'state': 'cancel'
+                        })
+                        move.unlink()
 
     @api.multi
     def _compute_serial_available(self):
